requirements.txt
accelerate==0.33.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
banal==1.0.6
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dataset==1.6.2
datasets==2.20.0
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.3
hydra-core==1.3.2
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.37.1
packaging==24.1
pandas==2.2.2
peft==0.12.0
pillow==10.4.0
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.53
sympy==1.13.1
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.43.3
triton==3.0.0
trl==0.9.6
typing_extensions==4.12.2
tyro==0.8.5
tzdata==2024.1
urllib3==2.2.2
xxhash==3.4.1
yarl==1.9.4
